+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   - Added parallax component
   - Added slideset component
   - Added slider component
+  - Added custom animation duration for kenburns effect (slideshow component)
   - Added support for overlaying an image with a second image to overlay component
   - Added trigger 'input' after htmleditor change
   - Added UIkit.init([rootelement]) added to manually init js components